+++ pypsrp/wsman.py
@@ -888,7 +888,7 @@ class _TransportHTTP(object):
     def _build_auth_kerberos(self, session):
         self._build_auth_negotiate(session, "kerberos")
 
-    def _build_auth_negotiate(self, session, auth_provider="auto"):
+    def _build_auth_negotiate(self, session, auth_provider="negotiate"):
         kwargs = self._get_auth_kwargs('negotiate')
 
         session.auth = HTTPNegotiateAuth(username=self.username,